Moved takeSnapshot into test-case-recorder package
