+++ packages/cursorless-engine/src/index.ts
@@ -1,5 +1,4 @@
 export * from "./testUtil/plainObjectToTarget";
-export * from "./testUtil/takeSnapshot";
 export * from "./core/StoredTargets";
 export * from "./cursorlessEngine";
 export * from "./generateSpokenForm/defaultSpokenForms/surroundingPairsDelimiters";
