+++ packages/test-case-recorder/src/TestCaseRecorder.ts
@@ -28,7 +28,6 @@ import {
   injectIde,
   SpokenFormGenerator,
   StoredTargetMap,
-  takeSnapshot,
 } from "@cursorless/cursorless-engine";
 import { getRecordedTestsDirPath, walkDirsSync } from "@cursorless/node-common";
 import { invariant } from "immutability-helper";
@@ -37,6 +36,7 @@ import * as fs from "node:fs";
 import { access, readFile } from "node:fs/promises";
 import * as path from "node:path";
 import { RecordTestCaseCommandOptions } from "./RecordTestCaseCommandOptions";
+import { takeSnapshot } from "./takeSnapshot";
 import { TestCase } from "./TestCase";
 
 const CALIBRATION_DISPLAY_DURATION_MS = 50;
