+++ packages/test-case-recorder/src/index.ts
@@ -1,2 +1,3 @@
-export * from "./TestCaseRecorder";
 export * from "./ScopeTestRecorder";
+export * from "./takeSnapshot";
+export * from "./TestCaseRecorder";
